requirements.txt
Flask==2.3.2
google-auth==2.24.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.96.0
python-dotenv==1.0.0

